Update patch set 2
Patch Set 2: Code-Review+1

> Patch Set 2:
> 
> These two tests started failing after the flag was flipped: https://ci.chromium.org/p/v8/builders/ci/V8%20Linux64%20GC%20Stress%20-%20custom%20snapshot/30797
> 
> They both test that a function is optimized, then deoptimized, then reoptimized taking new feedback into account. What I think is happening is that flipping the flag changed the GC pattern when stressing GC, and the new GC timing is unintentionally clearing type feedback at bad times, causing failures.
> 
> AFAIU, %PrepareFunctionForOptimization ensures a function's type feedback vector is not cleared on GC until the next %OptimizeFunctionOnNextCall. In these tests though, there's some mutator code that runs in between the an optimization event and the next %PrepareFunctionForOptimization, while the assumption seems to be that type feedback is kept for the entire duration of the opt-then-deopt-then-opt series of events. In other words, if the test is supposed to be opt-then-deopt-then-opt, %PrepareFunctionForOptimization must be called immediately after an optimization event to guard against untimely GCs.
> 
> I don't understand the deopt system nearly well enough to say for sure though. Ulan, does that make sense?

Thanks for debugging this. I think this is right. The PrepareFunctionForOptimization cause the bytecode array to be strongly held, which prevents it from being flushed, which would cause it to flush the type feedback as well (causing the steps you mention above). The bytecode can't be collected while the function is optimized, so theoretically it should be fine to avoid calling it until the function deopts (I thought PrepareFunctionForOptimization had a check where it wouldn't do anything if the function was already optimized, which would explain why it was after the deopt in the original code, but that doesn't seem to be the case), however it is indeed possible that some GC happens between a deopt and the call to PrepareFunctionForOptimization which could cause it to lose the type feedback, so moving it to where you have now sounds good.

LGTM, thanks.

Patch-set: 2
Reviewer: Gerrit User 1002763 <1002763@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1