Update patch set 31
Patch Set 31:

(11 comments)

Thanks, I really like the use of the heap. For committing you seem to need some updates in the end-to-end tests according to presubmit. Maybe a special end-to-end case for this feature would be also nice. Or maybe it'll be tested well enough with the test data we have.

Patch-set: 31
Reviewer: Gerrit User 1115956 <1115956@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>