Update patch set 3
Patch Set 3:

(2 comments)

Thanks for working on this!

I am not sure if the approach of storing a TaskRunner in the CompilationState is a good idea. The TaskRunner belongs to a single isolate, whereas NativeModules are shared between isolates. So once the first isolate dies, the TaskRunner will just drop all tasks and never execute them.

Do you think it would be possible to just never spawn any background tasks in single-threaded mode (also not as foreground tasks), and instead explicitly do the work "inline"? We would have to be careful then with surrounding scopes, but if we code this explicitly (instead of executing background tasks inline, as before the last CL), I think it should be OK.

Patch-set: 3
Reviewer: Gerrit User 1177599 <1177599@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>