[wasm-simd][arm] Improve codegen for any_true
The previous codegen used a bunch of vpmax to compare across lanes. We
can simplify a bit by checking if src is all 0. There is no way to
compare en entire SIMD register against 0, so we first OR the high and
low, then compare it with 0, and move 0 or 1 depending on the result.

The resulting codegen is 4 instructions (only 1 of the two movs will
happen), and works for all 3 kinds of any_true.

Change-Id: I66013796af08a666009e6b2d774ea7ee7bdfe1ad