Update patch set 6
Patch Set 6:

The codegen LGTM!

Out of curiosity, are those non-trapping operations performance critical? I'm wondering if it would be worth optimizing those in TurboFan, since at the moment we generate a `TruncateFloat32ToInt32()` op which lowers to:

```
    case kArmVcvtS32F32: {
      UseScratchRegisterScope temps(tasm());
      SwVfpRegister scratch = temps.AcquireS();
      __ vcvt_s32_f32(scratch, i.InputFloatRegister(0));
      __ vmov(i.OutputRegister(), scratch);
      // Avoid INT32_MAX as an overflow indicator and use INT32_MIN instead,
      // because INT32_MIN allows easier out-of-bounds detection.
      __ cmn(i.OutputRegister(), Operand(1));
      __ mov(i.OutputRegister(), Operand(INT32_MIN), SBit::LeaveCC, vs);
      DCHECK_EQ(LeaveCC, i.OutputSBit());
      break;
    }
```

And then we compare the result with INT32_MIN to check if we should trap. But if we're using a saturating opcode, then we could omit the `cmn` and `mov` instruction. I think the same goes for Arm64.

Thanks,
Pierre

Patch-set: 6
CC: Gerrit User 1183769 <1183769@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>