Make Math.random() not consume cache in no-side-effects mode
Math.random, while technically not having any effects which modify the
surrounding JS state, does observably change between a no-side-effects
evaluation and an actual evaluation, and can cause confusion. This CL
adds a check to MathRandom which causes it to continually use the same
cache entry if evaluated in no-side-effects mode.

Change-Id: I4a41ac6fd3153a14245d5940fe52ada43ca05e0b