Update patch set 1
Patch Set 1:

(1 comment)

Recording our conversation here:

For the mutability bit, my plan was to add a separate bit vector (std::vector<bool> should work) to the StructType class, as that's more memory efficient than an array of FieldType structs. It's hard to say whether that's premature optimization though, so we can also keep the FieldType if you prefer. There's not a whole lot of reason to exist for it though if you do the other thing we discussed, which is to add i8/i16 to the ValueType enum/class.

Patch-set: 1
Reviewer: Gerrit User 1115935 <1115935@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>