Update patch set 3
Patch Set 3:

> > Sure: test/wasm-api-tests/*. Our test driver can handle them like any other test suite.
> 
> I found these also, should I be running them? https://source.chromium.org/chromium/chromium/src/+/master:v8/tools/run-wasm-api-tests.py

That's a script to compile the upstream examples, link them against libwee8, and run them. For the type of changes you're doing, that shouldn't provide any additional coverage compared to the test/wasm-api-tests/* suite. (On the contrary; aside from compile/link failures and crashes, the upstream examples rely on human reading of stdout to detect misbehavior.)

Patch-set: 3