[WIP][sandbox][x64] Access external pointer through a table
This change moves external pointers into a separate table datastructure
and turns external pointers in heap objects into indices into that
table.

This CL implements one of two possible ownership models for the table
entries. With this one, a new table entry is allocated for every
external pointer stored in the v8 heap. As such, indices into the
external pointer table can be shared between different v8 heap objects.
Further, an entry in the table is only allocated when an external
pointer field is set. On the other hand, this approach requires more
logic in the garbage collector to detect and free stale table entries.
Also, it requires allocating a table entry every time an external
pointer is stored in a v8 heap object.

This CL does not yet implement memory management of the external pointer
table which will have to be done by the GC.

Bug: v8:10391
Change-Id: Ib78dcd90729dc3007e8eeeee9dea8cb509bf8f95