Update patch set 5
Patch Set 5:

(3 comments)

Thanks for tracking the issues on the failing tests down to this!

Changing the actual file in the third_party folder doesn't seem to be the best way. For one, the copyright notice might be confusing as there is no original copyright notice and it looks like it's reclaiming the copyright to V8. I think changes by the V8 authors should be made in a separate file. Furthermore, it will be hard to update if a newer test harness is required because it won't be obvious anymore what has been changed in the meantime.

Instead I would suggest keeping the test harness in the third_party folder as is and patch all the changes in a local test harness wrapper that loads this one and modifies it before returning it.

Patch-set: 5