[wasm][tests] Sync testharness of wasm-js tests with wpt
Recently the test harness of the js-api spec tests got updated.
To the test harness of wpt. With this CL we import the wpt test
harness into V8. The test harness required some updates though:

1) Changes to {promise_test}. In d8 unhandled exceptions in the
promise change simply get ignored. Therefore we adjusted the test
harness to catch unhandled exceptions, and to re-throw them in a
setTimeout to make the exception visible.

2) The js-types proposal did not get updated yet to use the new
test harness. Therefore we had to add the assert_throws and
promise_rejects functions to the test harness.

The original test harness of wpt can be seen in PS 2. Later
patchsets then contain the changes to the test harness.

Bug: v8:10556
Change-Id: Ia5f7dec3f40ba8fc639135fb2b2078b9220eccf7