Update patch set 2
Patch Set 2:

isn't the right behavior that they aren't marked as maybe_assigned though? The value doesn't change after the declaration. Is this too hard to get right in the preparser? (To be honest it doesn't seem super important to me, but it would be a shame for it to be missing out on the optimization.)

Patch-set: 2