Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> isn't the right behavior that they aren't marked as maybe_assigned though? The value doesn't change after the declaration. Is this too hard to get right in the preparser? (To be honest it doesn't seem super important to me, but it would be a shame for it to be missing out on the optimization.)

Yes probably. I was taking the view that this wasn't something to spend lots of time optimising for. On the other hand I'm concerned that my current solution may impose a cost on more compliant code.

I'll look into changing the path that adds the variable to not mark them as assigned and see if that works.

Patch-set: 3