Update patch set 3
Patch Set 3:

I think you misunderstood this github issue. Anyfunc is not a string that we picked in V8 and now is outdated due to spec changes. The string "anyfunc" is actually part of the spec, see e.g. https://webassembly.github.io/spec/js-api/index.html#tables. If you read through the github issue Jakob mentioned (https://github.com/WebAssembly/reference-types/pull/85), it actually says that we cannot rename anyfunc to funcref. They propose an alias, but this did not happen. I think if you run the spec tests, you would actually say that with this change we would not pass it.

Patch-set: 3