Update patch set 4
Patch Set 4:

This is a CL to add event logging in ETW for Windows and Signposts for MacOSX. They share a common API that can be filled in for other OSs as well.

This is my first CL and I don't have TryBot or TryJob access yet. As such, I haven't been able to do a dry-run yet. Can any of you guys help me with that? Thanks!

Patch-set: 4
Label: Commit-Queue=0
Work-in-progress: false