[wasm] Speed up ValueType::element_size_bytes
It turns out that Liftoff often needs to know the size of a value in
bytes. Currently we are loading the size_log_2 from an array and then
performing a shift by that amount. We can slightly speed this up by just
loading the correct value directly.

Drive-by: Use {int8_t} for the internal array, since all values will
easily fit in that range.

R=thibaudm@chromium.org

Bug: v8:10576
Change-Id: I1b832ba404ff9913e2272d332f312b371b6ce3d4