[wasm] Properly implement parsing of s33 values
We used to approximate s33 value parsing by first checking for specific
negative codes, and then parsing an u32 value if that failed. This is
not correct in all cases. This CL implements s33 parsing properly and
uses it where the wasm spec requires it.
Additionally, parsing of heap types has been factored out into a macro.

Change-Id: I04deabce8837a48af2226411cd706a397f9e5725