[wasm] Untangle decoding of different opcodes
This CL mostly removes code paths which join again after a big switch.
Instead of updating a local variable (for the length of the decoded
operation), and then breaking out of the switch just to return the value
of that local variable, we now just return directly from within the
switch.
This will allow to
1) split the switch into individual methods per opcode, and
2) tail-call to helpers like {BuildSimpleOperator} or
   {DecodeNumericOpcode}.

R=thibaudm@chromium.org

Bug: v8:10576
Change-Id: I466068d3566c6bff8bd3ac78a2bba60c3663dd52